move storage types to their own files
