+++ crates/batch-prover/src/proving.rs
@@ -12,7 +12,8 @@ use prover_services::{ParallelProverService, ProofData};
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
 use sov_db::ledger_db::BatchProverLedgerOps;
-use sov_db::schema::types::{SoftConfirmationNumber, StoredBatchProof, StoredBatchProofOutput};
+use sov_db::schema::types::batch_proof::{StoredBatchProof, StoredBatchProofOutput};
+use sov_db::schema::types::SoftConfirmationNumber;
 use sov_modules_api::{SlotData, SpecId, Zkvm};
 use sov_rollup_interface::da::{BlockHeaderTrait, DaNamespace, DaSpec, SequencerCommitment};
 use sov_rollup_interface::rpc::SoftConfirmationStatus;
