+++ crates/light-client-prover/src/da_block_handler.rs
@@ -9,7 +9,8 @@ use citrea_primitives::forks::fork_from_block_number;
 use prover_services::{ParallelProverService, ProofData};
 use sov_db::ledger_db::{LightClientProverLedgerOps, SharedLedgerOps};
 use sov_db::mmr_db::MmrDB;
-use sov_db::schema::types::{SlotNumber, StoredLightClientProofOutput};
+use sov_db::schema::types::light_client_proof::StoredLightClientProofOutput;
+use sov_db::schema::types::SlotNumber;
 use sov_modules_api::{BatchProofCircuitOutputV2, BlobReaderTrait, DaSpec, Zkvm};
 use sov_rollup_interface::da::{BlockHeaderTrait, DaDataLightClient, DaNamespace};
 use sov_rollup_interface::mmr::{MMRChunk, MMRNative, Wtxid};
